[IMP] web: calendar: batch writes on filters
This commit improves the way calendar filters are written in
database. In the calendar view, a filter with the `write_model` and
`filter_field` attributes has its filter values and status stored
in database. It means that when a user ticks/unticks such a
filter's value in the filter panel, the new status is saved in
database. Before this commit, ticking/unticking all filters via
the main checkbox displayed next to the section name performed as
many write rpcs as the number of values in the section. With this
commit, a single write is done for all records.

This commit also improves the fetching of the filters. In the
hypothetical case where there're more than one filter with its
dedicated model (i.e. that needs to be read from the server),
those filters were read sequentially. They're now read in parallel.

Task-4567282

Part-of: odoo/odoo#197338
Related: odoo/enterprise#79244
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>